requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
Automat==22.10.0
autoscraper==1.1.14
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.41.0
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
Protego==0.3.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
Scrapy==2.11.2
selenium==4.23.1
service-identity==24.1.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.1
tldextract==5.1.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
tweepy==4.14.0
Twisted==24.7.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
w3lib==2.2.1
websocket-client==1.8.0
Werkzeug==2.2.3
whitenoise==6.7.0
wsproto==1.2.0
zope.interface==7.0.1
